split Vast parser into 2 classes - one only for parsing one XML file. Second that recursively unwraps any wrapper response it might receive.
