+++ VastClient/VastClient/constants/VastMacros.swift
@@ -10,7 +10,8 @@ import Foundation
 
 enum VatMacros: String {
     case errorcode = "[ERRORCODE]"
-    case contentplayhead = "[CONTENTPLAYHEAD]"
-    case cachebusting = "[CACHEBUSTING]"
+    case contentplayhead = "[CONTENTPLAYHEAD]" //current time offset “HH:MM:SS.mmm” of the video content
+    case cachebusting = "[CACHEBUSTING]" //replaced with a random 8-digit number.
     case asseturi = "[ASSETURI]"
+    case timestapm = "[TIMESTAMP]" //date and time at which the URI using this macro is accessed using ISO 8601 with miliseconds: January 17, 2016 at 8:15:07 and 127 milleseconds, Eastern Time would be formatted as follows: 2016-01-17T8:15:07.127-05
 }
