+++ VastClient/VastClientTests/Assets/Vast3/InlineLinearTag-test3.swift
@@ -45,7 +45,7 @@ extension VastModel {
             // TODO: this is not part of the current model
         ]
         
-        let ad = VastAd(type: .inline, id: "20001", sequence: 0, conditionalAd: nil, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", wrapperUrl: nil, adCategories: [], description: nil, advertiser: nil, surveys: [], wrapper: nil)
+        let ad = VastAd(type: .inline, id: "20001", sequence: 0, conditionalAd: nil, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", adCategories: [], description: nil, advertiser: nil, surveys: [], wrapper: nil)
     
         return VastModel(version: "3.0", ads: [ad], errors: [])
     }()
