+++ VastClient/VastClientTests/Assets/Vast4/Category-test.swift
@@ -50,7 +50,7 @@ extension VastModel {
         
         let extensions: [VastExtension] = []
         
-        let ad = VastAd(type: .inline, id: "20008", sequence: 1, conditionalAd: false, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", wrapperUrl: nil, adCategories: categories, description: nil, advertiser: nil, surveys: [], wrapper: nil)
+        let ad = VastAd(type: .inline, id: "20008", sequence: 1, conditionalAd: false, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", adCategories: categories, description: nil, advertiser: nil, surveys: [], wrapper: nil)
         return VastModel(version: "4.0", ads: [ad], errors: [])
     }()
 }
