+++ VastClient/VastClientTests/Assets/Vast4/NoWrapperTag-test.swift
@@ -48,7 +48,7 @@ extension VastModel {
             VastExtension(type: "iab-Count", creativeParameters: [])
         ]
         
-        let ad = VastAd(type: .inline, id: "20001", sequence: 1, conditionalAd: false, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", wrapperUrl: nil, adCategories: categories, description: nil, advertiser: nil, surveys: [], wrapper: nil)
+        let ad = VastAd(type: .inline, id: "20001", sequence: 1, conditionalAd: false, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", adCategories: categories, description: nil, advertiser: nil, surveys: [], wrapper: nil)
         
         return VastModel(version: "4.0", ads: [ad], errors: [])
     }()
