+++ VastClient/VastClientTests/Assets/Vast4/VideoClicksAndClickTracking-Inline-test.swift
@@ -51,7 +51,7 @@ extension VastModel {
             VastExtension(type: "iab-Count", creativeParameters: [])
         ]
         
-        let ad = VastAd(type: .inline, id: "20009", sequence: 1, conditionalAd: false, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", wrapperUrl: nil, adCategories: categories, description: nil, advertiser: nil, surveys: [], wrapper: nil)
+        let ad = VastAd(type: .inline, id: "20009", sequence: 1, conditionalAd: false, adSystem: adSystem, impressions: [impressions], adVerifications: [], viewableImpression: nil, pricing: pricing, errors: errors, creatives: [creative], extensions: extensions, adTitle: "iabtechlab video ad", adCategories: categories, description: nil, advertiser: nil, surveys: [], wrapper: nil)
         
         return VastModel(version: "4.0", ads: [ad], errors: [])
     }()
